% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxCoxinv.R
\name{BoxCox.inv}
\alias{BoxCox.inv}
\title{Inverse Box-cox transform a vector}
\usage{
BoxCox.inv(x, lambda, plot.it = F, omit.na = T)
}
\arguments{
\item{x}{The vector to be transformed.}

\item{lambda}{The index that is used to do Box-Cox inverse transformation.}

\item{plot.it}{logical param. True if want \code{x} against transformed results to be plotted}

\item{omit.na}{logical param. True if want NA in \code{x} to be omited.}
}
\value{
A vector that is the Box-Cox inverse transformation of \code{x}.
}
\description{
That's it -- this function just transform a vector by Box-Cox inverses.
}
\details{
This function may need positive values of x for certain lambda.

Please see error message for more information.

Here are some reasons why putting a list in this section is excessive:
\itemize{
     \item This \code{BoxCox.inv} function is quite simple.
     \item There's nothing else to say about \code{BoxCox.inv}.
}
}
\examples{
BoxCox.inv(1:10,lambda=2)
BoxCox.inv(-5,lambda=-1)
}
