% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxCox.R
\name{BoxCox}
\alias{BoxCox}
\title{Box-cox transform a vector}
\usage{
BoxCox(x, lambda, plot.it = F, omit.na = T)
}
\arguments{
\item{x}{The vector to be transformed.}

\item{lambda}{The index that is used to do Box-Cox transformation.}

\item{plot.it}{logical param. True if want \code{x} against transformed results to be plotted}

\item{omit.na}{logical param. True if want NA in \code{x} to be omited.}
}
\value{
A vector that is the Box-Cox transformation of \code{x}.
}
\description{
That's it -- this function just transform a vector by Box-Cox.
}
\details{
This function may need positive values of x for certain lambda.

Please see error message for more information.

Here are some reasons why putting a list in this section is excessive:
\itemize{
     \item This \code{BoxCox} function is quite simple.
     \item There's nothing else to say about \code{BoxCox}.
}
}
\examples{
BoxCox(1:10,lambda=2)
BoxCox(-5,lambda=-1)
}
