% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{cube}
\alias{cube}
\title{Cube a vector}
\usage{
cube(x, plot.it = FALSE)
}
\arguments{
\item{x}{The vector to be cubed.}
}
\value{
A vector that is the cube of \code{x}.
}
\description{
That's it -- this function just cube a vector.
}
\details{
Inorder to build a package first load package roxygen2

use roxygen2::roxygenize()


Here are some reasons why putting a list in this section is excessive:
\itemize{
     \item This \code{cube} function is quite simple.
     \item There's nothing else to say about \code{cube}.
}
}
\examples{
cube(1:10)
cube(-5)
}
